ai_service/requirements.txt
flask
python-docx
numpy
scikit-learn
google-generativeai 
google.genai
dotenv